reuse size_t instead of uint16_t minimize binary size (templates)

+++ libmamba/include/mamba/core/context.hpp
@@ -130,7 +130,7 @@ namespace mamba
         ChannelPriority channel_priority = ChannelPriority::kFlexible;
         bool auto_activate_base = false;
 
-        std::uint16_t download_threads = 5;
+        std::size_t download_threads = 5;
         int extract_threads = 0;
 
         int verbosity = 0;
